requirements.txt
langgraph>=1.0.1
haystack-ai>=2.0.0
openai>=1.0.0
qdrant-client>=1.6.0
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
numpy>=1.24.0
pandas>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
ragas>=0.0.22
langfuse>=2.0.0
grpcio>=1.50.0
grpcio-tools>=1.50.0